Ad a dashboard BS theme and move js scripts to modules
