+++ shedpi_hub_dashboard/static/shedpi_hub_dashboard/js/modules/utils.js
@@ -1,0 +1,12 @@
+/* Shared utils */
+function getSchemaDataFields(schema) {
+  let dataFields = []
+  if (schema) {
+    let extra_fields = Object.keys(schema.properties)
+    dataFields = [...dataFields, ...extra_fields];
+    dataFields = [...new Set(dataFields)]
+  }
+  return dataFields
+}
+
+export {getSchemaDataFields}
