+++ shedpi_hub_dashboard/templates/shedpi_hub_dashboard/index.html
@@ -13,44 +13,174 @@
 
 </head>
 <body>
-<header>
-  <h1>Shed Pi data</h1>
-</header>
-<div class="contents" data-endpoint-url="">
 
-  <form class="row g-3">
+<header class="navbar sticky-top bg-dark flex-md-nowrap p-0 shadow" data-bs-theme="dark">
+  <a class="navbar-brand col-md-3 col-lg-2 me-0 px-3 fs-6 text-white" href="#">ShedPi</a>
 
-    <div class="col-lg-3 col-sm-6">
-      <label for="startDate">Start</label>
-      <input id="startDate" class="form-control" type="date"/>
-      <span id="startDateSelected"></span>
-    </div>
-    <div class="col-lg-3 col-sm-6">
-      <label for="endDate">End</label>
-      <input id="endDate" class="form-control" type="date"/>
-      <span id="endDateSelected"></span>
-    </div>
-    <div class="col-lg-3 col-sm-6">
-      <select class="form-select" id="form-dropdown" aria-label="Device module selector">
-      </select>
-    </div>
+  <ul class="navbar-nav flex-row d-md-none">
+    <li class="nav-item text-nowrap">
+      <button class="nav-link px-3 text-white" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSearch"
+              aria-controls="navbarSearch" aria-expanded="false" aria-label="Toggle search">
+        <svg class="bi">
+          <use xlink:href="#search"/>
+        </svg>
+      </button>
+    </li>
+    <li class="nav-item text-nowrap">
+      <button class="nav-link px-3 text-white" type="button" data-bs-toggle="offcanvas" data-bs-target="#sidebarMenu"
+              aria-controls="sidebarMenu" aria-expanded="false" aria-label="Toggle navigation">
+        <svg class="bi">
+          <use xlink:href="#list"/>
+        </svg>
+      </button>
+    </li>
+  </ul>
 
-    <div class="col-lg-3 col-sm-6">
-      <button type="submit" class="btn btn-primary mb-3" id="form-submit">Submit</button>
-    </div>
-
-  </form>
-  <div class="row">
-    <div class="col">
-      <div id="chartContainer"></div>
-    </div>
+  <div id="navbarSearch" class="navbar-search w-100 collapse">
+    <input class="form-control w-100 rounded-0 border-0" type="text" placeholder="Search" aria-label="Search">
   </div>
+</header>
 
+<div class="contents container-fluid">
+  <div class="row">
+    <div class="sidebar border border-right col-md-3 col-lg-2 p-0 bg-body-tertiary">
+      <div class="offcanvas-md offcanvas-end bg-body-tertiary" tabindex="-1" id="sidebarMenu"
+           aria-labelledby="sidebarMenuLabel">
+        <div class="offcanvas-header">
+          <h5 class="offcanvas-title" id="sidebarMenuLabel">Company name</h5>
+          <button type="button" class="btn-close" data-bs-dismiss="offcanvas" data-bs-target="#sidebarMenu"
+                  aria-label="Close"></button>
+        </div>
+        <div class="offcanvas-body d-md-flex flex-column p-0 pt-lg-3 overflow-y-auto">
+          <ul class="nav flex-column">
+            <li class="nav-item">
+              <a class="nav-link d-flex align-items-center gap-2 active" aria-current="page" href="#">
+                <svg class="bi">
+                  <use xlink:href="#"/>
+                </svg>
+                Dashboard
+              </a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link d-flex align-items-center gap-2" href="#">
+                <svg class="bi">
+                  <use xlink:href="#"/>
+                </svg>
+                Other
+              </a>
+            </li>
+          </ul>
+
+          <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-body-secondary text-uppercase">
+            <span>Saved reports</span>
+            <a class="link-secondary" href="#" aria-label="Add a new report">
+              <svg class="bi">
+                <use xlink:href="#plus-circle"/>
+              </svg>
+            </a>
+          </h6>
+          <ul class="nav flex-column mb-auto">
+            <li class="nav-item">
+              <a class="nav-link d-flex align-items-center gap-2" href="#">
+                <svg class="bi">
+                  <use xlink:href="#file-earmark-text"/>
+                </svg>
+                Current month
+              </a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link d-flex align-items-center gap-2" href="#">
+                <svg class="bi">
+                  <use xlink:href="#file-earmark-text"/>
+                </svg>
+                Last quarter
+              </a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link d-flex align-items-center gap-2" href="#">
+                <svg class="bi">
+                  <use xlink:href="#file-earmark-text"/>
+                </svg>
+                Social engagement
+              </a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link d-flex align-items-center gap-2" href="#">
+                <svg class="bi">
+                  <use xlink:href="#file-earmark-text"/>
+                </svg>
+                Year-end sale
+              </a>
+            </li>
+          </ul>
+
+          <hr class="my-3">
+
+          <ul class="nav flex-column mb-auto">
+            <li class="nav-item">
+              <a class="nav-link d-flex align-items-center gap-2" href="#">
+                <svg class="bi">
+                  <use xlink:href="#gear-wide-connected"/>
+                </svg>
+                Settings
+              </a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link d-flex align-items-center gap-2" href="#">
+                <svg class="bi">
+                  <use xlink:href="#door-closed"/>
+                </svg>
+                Sign out
+              </a>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+
+    <main class="col-md-9 ms-sm-auto col-lg-10 px-md-4">
+      <div
+          class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
+        <h1 class="h2">Dashboard</h1>
+      </div>
+
+      <form class="row g-3">
+
+        <div class="col-lg-3 col-sm-6">
+          <label for="startDate">Start</label>
+          <input id="startDate" class="form-control" type="date"/>
+          <span id="startDateSelected"></span>
+        </div>
+        <div class="col-lg-3 col-sm-6">
+          <label for="endDate">End</label>
+          <input id="endDate" class="form-control" type="date"/>
+          <span id="endDateSelected"></span>
+        </div>
+        <div class="col-lg-3 col-sm-6">
+          <select class="form-select" id="form-dropdown" aria-label="Device module selector">
+          </select>
+        </div>
+
+        <div class="col-lg-3 col-sm-6">
+          <button type="submit" class="btn btn-primary mb-3" id="form-submit">Submit</button>
+        </div>
+
+      </form>
+
+      <h2>Chart</h2>
+      <div id="chartContainer"></div>
+
+      <h2>Section title</h2>
+      <div id="tableContainer"></div>
+
+    </main>
+  </div>
 </div>
+
 <footer>
 </footer>
 <script src="https://cdn.jsdelivr.net/npm/d3@7"></script>
-<script src="{% static 'shedpi_hub_dashboard/js/index.js' %}"></script>
+<script src="{% static 'shedpi_hub_dashboard/js/index.js' %}" type="module"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
         crossorigin="anonymous"></script>
